refactor: relocate mock_fake_clock for use in rest library
