+++ google/cloud/storage/storage_client_testing.bzl
@@ -21,7 +21,6 @@ storage_client_testing_hdrs = [
     "testing/client_unit_test.h",
     "testing/constants.h",
     "testing/mock_client.h",
-    "testing/mock_fake_clock.h",
     "testing/mock_http_request.h",
     "testing/mock_storage_stub.h",
     "testing/object_integration_test.h",
@@ -35,7 +34,6 @@ storage_client_testing_hdrs = [
 
 storage_client_testing_srcs = [
     "testing/client_unit_test.cc",
-    "testing/mock_fake_clock.cc",
     "testing/mock_http_request.cc",
     "testing/object_integration_test.cc",
     "testing/random_names.cc",
