+++ google/cloud/testing_util/google_cloud_cpp_testing.bzl
@@ -27,6 +27,7 @@ google_cloud_cpp_testing_hdrs = [
     "expect_exception.h",
     "expect_future_error.h",
     "integration_test.h",
+    "mock_fake_clock.h",
     "scoped_environment.h",
     "scoped_log.h",
     "scoped_thread.h",
@@ -40,6 +41,7 @@ google_cloud_cpp_testing_srcs = [
     "crash_handler.cc",
     "example_driver.cc",
     "integration_test.cc",
+    "mock_fake_clock.cc",
     "scoped_environment.cc",
     "scoped_log.cc",
     "status_matchers.cc",
