+++ google/cloud/testing_util/mock_fake_clock.cc
@@ -1,4 +1,4 @@
-// Copyright 2020 Google LLC
+// Copyright 2021 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,18 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/testing/mock_fake_clock.h"
+#include "google/cloud/testing_util/mock_fake_clock.h"
 
 namespace google {
 namespace cloud {
-namespace storage {
-namespace testing {
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
+namespace testing_util {
 
 // We need to define this value since it is static; set it to an arbitrary
 // value.
 std::time_t FakeClock::now_value_ = 1530060324;
 
-}  // namespace testing
-}  // namespace storage
+}  // namespace testing_util
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace cloud
 }  // namespace google
